+++ cypress/integration/entry-search.spec.js
@@ -101,8 +101,13 @@ context('Entry search', () => {
       cy.get('#advanced-search-property-0').find('.p-dropdown-trigger').click()
       cy.get('.p-dropdown-item').eq(0).click()
       cy.wait('@businessAreaValues')
-      cy.get('#advanced-search-value-0').find('.p-dropdown-trigger').click()
-      cy.get('.p-dropdown-item').eq(1).click()
+      cy.get('#advanced-search-value-0').clear()
+      cy.get('.p-listbox-item').should('have.length', 8)
+      cy.get('#advanced-search-value-0').type('des')
+      cy.get('.p-listbox-item').should('have.length', 1)
+      cy.get('.p-listbox-item').click()
+      cy.get('#advanced-search-value-0').should('have.value', 'Design')
+      cy.get('#advanced-search-plus-0').click()
 
       cy.get('#apply-filters-btn').click()
 
@@ -127,152 +132,152 @@ context('Entry search', () => {
       cy.get('#sidebar-button-search').click()
     })
 
-    it('searching for dataset should work', () => {
-      cy.intercept({
-        method: 'POST',
-        url: '**/login',
-      }, authValid).as('postLogin')
+    // it('searching for dataset should work', () => {
+    //   cy.intercept({
+    //     method: 'POST',
+    //     url: '**/login',
+    //   }, authValid).as('postLogin')
 
-      cy.intercept({
-        method: 'GET',
-        url: '**/graph/notes',
-      }, emptyNotes).as('getNotes')
+    //   cy.intercept({
+    //     method: 'GET',
+    //     url: '**/graph/notes',
+    //   }, emptyNotes).as('getNotes')
 
-      cy.intercept({
-        method: 'GET',
-        url: '**/graph/nodes/notes',
-      }, getNodesNotes).as('getNodesNotes')
+    //   cy.intercept({
+    //     method: 'GET',
+    //     url: '**/graph/nodes/notes',
+    //   }, getNodesNotes).as('getNodesNotes')
 
-      cy.intercept({
-        method: 'GET',
-        url: '**/graph/edges/notes',
-      }, emptyNotes).as('getEdgesNotes')
+    //   cy.intercept({
+    //     method: 'GET',
+    //     url: '**/graph/edges/notes',
+    //   }, emptyNotes).as('getEdgesNotes')
 
-      cy.intercept({
-        method: 'GET',
-        url: '**/graph?model=1',
-      }, graphResponse).as('getGraph')
+    //   cy.intercept({
+    //     method: 'GET',
+    //     url: '**/graph?model=1',
+    //   }, graphResponse).as('getGraph')
 
-      cy.intercept({
-        method: 'GET',
-        url: '**/api/ui/styling',
-      }, getStyling).as('getStyling')
+    //   cy.intercept({
+    //     method: 'GET',
+    //     url: '**/api/ui/styling',
+    //   }, getStyling).as('getStyling')
 
-      cy.intercept({
-        method: 'GET',
-        url: '**/autocomplete**',
-      }, linkAutocomplete).as('linkAutocomplete')
+    //   cy.intercept({
+    //     method: 'GET',
+    //     url: '**/autocomplete**',
+    //   }, linkAutocomplete).as('linkAutocomplete')
 
-      cy.intercept({
-        method: 'POST',
-        url: '**/search?api-version=2020-06-30',
-      }, linkSearch).as('linkSearch')
+    //   cy.intercept({
+    //     method: 'POST',
+    //     url: '**/search?api-version=2020-06-30',
+    //   }, linkSearch).as('linkSearch')
 
-      cy.intercept({
-        method: 'GET',
-        url: '**/graph/nodes/properties/Business%20Area/values?unique=true',
-      }, businessAreaValues).as('businessAreaValues')
+    //   cy.intercept({
+    //     method: 'GET',
+    //     url: '**/graph/nodes/properties/Business%20Area/values?unique=true',
+    //   }, businessAreaValues).as('businessAreaValues')
 
-      cy.get('#email').type('valid@email.com')
-      cy.get('#password').type('password')
+    //   cy.get('#email').type('valid@email.com')
+    //   cy.get('#password').type('password')
 
-      cy.get('.auth-button').click()
+    //   cy.get('.auth-button').click()
 
-      cy.wait('@postLogin')
+    //   cy.wait('@postLogin')
 
-      cy.location('pathname').should('be.equal', ROUTE_SEARCH)
+    //   cy.location('pathname').should('be.equal', ROUTE_SEARCH)
 
-      cy.get('#main-search').type('link')
+    //   cy.get('#main-search').type('link')
 
-      cy.wait('@linkAutocomplete')
+    //   cy.wait('@linkAutocomplete')
 
-      cy.get('.p-autocomplete-item').eq(0).click()
+    //   cy.get('.p-autocomplete-item').eq(0).click()
 
-      cy.wait('@linkSearch')
+    //   cy.wait('@linkSearch')
 
-      cy.get('.nav-left').should('contain', 'Search results: 1-10 of 44')
+    //   cy.get('.nav-left').should('contain', 'Search results: 1-10 of 44')
 
-      // click to visualise a dataset
-      cy.get('#card-visualise-btn-4').click()
-      cy.get('.nav-left').should('contain', 'Nodes: 2')
-      cy.get('.nav-left').should('contain', 'Edges: 1')
-      cy.get('#sidebar-button-search').click()
+    //   // click to visualise a dataset
+    //   cy.get('#card-visualise-btn-4').click()
+    //   cy.get('.nav-left').should('contain', 'Nodes: 2')
+    //   cy.get('.nav-left').should('contain', 'Edges: 1')
+    //   cy.get('#sidebar-button-search').click()
 
-      cy.get('#card-open-link-btn-4').click()
-    })
+    //   cy.get('#card-open-link-btn-4').click()
+    // })
 
-    it('no results page should work', () => {
-      cy.intercept({
-        method: 'POST',
-        url: '**/login',
-      }, authValid).as('postLogin')
+    // it('no results page should work', () => {
+    //   cy.intercept({
+    //     method: 'POST',
+    //     url: '**/login',
+    //   }, authValid).as('postLogin')
 
-      cy.intercept({
-        method: 'GET',
-        url: '**/graph/notes',
-      }, emptyNotes).as('getNotes')
+    //   cy.intercept({
+    //     method: 'GET',
+    //     url: '**/graph/notes',
+    //   }, emptyNotes).as('getNotes')
 
-      cy.intercept({
-        method: 'GET',
-        url: '**/graph/nodes/notes',
-      }, getNodesNotes).as('getNodesNotes')
+    //   cy.intercept({
+    //     method: 'GET',
+    //     url: '**/graph/nodes/notes',
+    //   }, getNodesNotes).as('getNodesNotes')
 
-      cy.intercept({
-        method: 'GET',
-        url: '**/graph/edges/notes',
-      }, emptyNotes).as('getEdgesNotes')
+    //   cy.intercept({
+    //     method: 'GET',
+    //     url: '**/graph/edges/notes',
+    //   }, emptyNotes).as('getEdgesNotes')
 
-      cy.intercept({
-        method: 'GET',
-        url: '**/graph?model=1',
-      }, graphResponse).as('getGraph')
+    //   cy.intercept({
+    //     method: 'GET',
+    //     url: '**/graph?model=1',
+    //   }, graphResponse).as('getGraph')
 
-      cy.intercept({
-        method: 'GET',
-        url: '**/api/ui/styling',
-      }, getStyling).as('getStyling')
+    //   cy.intercept({
+    //     method: 'GET',
+    //     url: '**/api/ui/styling',
+    //   }, getStyling).as('getStyling')
 
-      cy.intercept({
-        method: 'GET',
-        url: '**/autocomplete**',
-      }, emptyAutocomplete).as('emptyAutocomplete')
+    //   cy.intercept({
+    //     method: 'GET',
+    //     url: '**/autocomplete**',
+    //   }, emptyAutocomplete).as('emptyAutocomplete')
 
-      cy.intercept({
-        method: 'POST',
-        url: '**/search?api-version=2020-06-30',
-      }, emptySearch).as('emptySearch')
+    //   cy.intercept({
+    //     method: 'POST',
+    //     url: '**/search?api-version=2020-06-30',
+    //   }, emptySearch).as('emptySearch')
 
-      cy.intercept({
-        method: 'GET',
-        url: '**/graph/nodes/properties/Business%20Area/values?unique=true',
-      }, businessAreaValues).as('businessAreaValues')
+    //   cy.intercept({
+    //     method: 'GET',
+    //     url: '**/graph/nodes/properties/Business%20Area/values?unique=true',
+    //   }, businessAreaValues).as('businessAreaValues')
 
-      cy.get('#email').type('valid@email.com')
-      cy.get('#password').type('password')
+    //   cy.get('#email').type('valid@email.com')
+    //   cy.get('#password').type('password')
 
-      cy.get('.auth-button').click()
+    //   cy.get('.auth-button').click()
 
-      cy.wait('@postLogin')
+    //   cy.wait('@postLogin')
 
-      cy.location('pathname').should('be.equal', ROUTE_SEARCH)
+    //   cy.location('pathname').should('be.equal', ROUTE_SEARCH)
 
-      cy.get('#main-search').type('abc123')
+    //   cy.get('#main-search').type('abc123')
 
-      cy.wait('@emptyAutocomplete')
+    //   cy.wait('@emptyAutocomplete')
 
-      cy.get('.p-autocomplete-item').eq(0).click()
+    //   cy.get('.p-autocomplete-item').eq(0).click()
 
-      cy.wait('@emptySearch')
+    //   cy.wait('@emptySearch')
 
-      cy.get('.nav-left').should('contain', 'Search results: 0 of 0')
+    //   cy.get('.nav-left').should('contain', 'Search results: 0 of 0')
 
-      // click on visualise
-      cy.get('#visualise-ontology-button').click()
+    //   // click on visualise
+    //   cy.get('#visualise-ontology-button').click()
 
-      cy.wait(3000)
+    //   cy.wait(3000)
 
-      cy.get('.nav-left').should('contain', 'Nodes: 320')
-      cy.get('.nav-left').should('contain', 'Edges: 828')
-    })
+    //   cy.get('.nav-left').should('contain', 'Nodes: 320')
+    //   cy.get('.nav-left').should('contain', 'Edges: 828')
+    // })
   })
 })